Use `kustomize edit fix` to update resources.
Added the `Kustomization` kind and `kustomize.config.k8s.io/v1beta1` apiVersion to all kustomization.yaml files in various config folders. This change serves to declare what the `kind` of API object the configuration file will construct and the `apiVersion` to interpret the `kind` with, which is a requirement for Kustomize configuration files. Additionally, moved from `bases` to `resources` field in the default config to align with updated kustomize usage rules.

Also modified `commonLabels` to `labels` field and added `includeSelectors` to ensure application metadata is correctly propagated to all Kubernetes objects related to the Pomerium application, including in the `stress-test` namespace.