more validation/status implementation
Fix the Listener allowedRefs processing. We need to fetch Namespace
objects in order to evaluate the labels (I had it originally validating
against the route labels, not the namespaces labels, which was
incorrect.) Add a Watches() call for namespace objects in the controller
setup.

Start to implement HTTPRoute validation. We need to fetch Service
objects for this, to be able to validate backendRefs. Add a
corresponding Watches() call for this as well.

Conformance test suite results: 29 passed, 20 failed, 44 skipped.