start to implement ReferenceGrant checks
Add a type for storing ReferenceGrant relationships as a map, for ease
of lookup. Rework the fetch logic to use this new type, and consult the
map when resolving route backendRefs.

Still need to update the actual translation logic to discard any invalid
backendRefs when converting to a Pomerium config proto.

Add a RouteDirectResponse to serve a 500 if there are no "To" URLs after
translating backendRefs.

Add temporary logic to switch on the insecure_server setting if there
are no HTTPS listeners. (This should let us get some value out of the
HTTP-only conformance tests, at least for now.)