better http/https switching hack, few fixes
Improve the hack for setting insecure_server so that it considers not
only the set of available listeners but the actual route attachments.
This should allow more conformance tests to pass, but I don't think it's
a viable option for when we clean this up to start merging things.

Fix how the HTTPRoute status entries are populated to handle the case
that a route specifies the same Gateway in two different parentRefs.

Ignore HTTPRouteMatch items with any options that aren't supported yet
(this improves the pass rate on the HTTPRouteMatchingAcrossRoutes test
cases).