pomerium: Fix Enpoints port mapping (#1102)
* pomerium: Add tests for Service to Endpoints mapping

Test more combinations of named and numbered ports, and in particular
avoid tests that happen to use the same name for the `port` and
`targetPort`, as they could pass despite the implementation being
broken.

* pomerium: Fix Enpoints port mapping

When trying to determine which Endpoints port to use based on a named
Service port, we weren't finding the right one when the Service
`spec.ports[].name` didn't match the Service `spec.ports[].targetPort`
name.

This commit simplifies the mapping logic, and ensure the extended test
from the previous commit all succeed.

Signed-off-by: Benoît Knecht <bknecht@protonmail.ch>